requirements.txt
robotframework==5.0.1
robotframework-appiumlibrary==2.0.0
Appium-Python-Client==5.1.2
robotframework-seleniumlibrary==6.0.0
requests==2.32.0
pyyaml==6.0.2
robotframework-requests==0.9.6
axe-selenium-python==2.1.6
robotframework-axelibrary==0.1.5